Unify kubeflow/local integration tests
